+++ src/auth/auth.controller.ts
@@ -12,8 +12,8 @@ import JwtOrApiKeyAuthGuard from './guards/jwt-apiKey-auth.guard'
 import { JwtAuthGuard } from './guards/jwt-auth.guard'
 import { LocalAuthGuard } from './guards/local-auth.guard'
 import { AuthService } from './auth.service'
+import { User } from '../user/decorators/user.decorator'
 import { User as UserEntity } from '../user/entities/user.entity'
-import { User } from '../user/user.decorator'
 import LoginDto from './dto/login.dto'
 import { SignUpDto } from './dto/signup.dto'
 import TokenResponseDto from './dto/token-response.dto'
